requirements.txt
telepot==12.7
PyYAML==6.0